[♻️ REFACTOR] : Menu 선택 코드 분리
